organize DatetimeIndex tests
